data-portal/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.13.0
stripe>=7.0.0
fastmcp>=3.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
PyJWT[crypto]>=2.8.0
gunicorn>=22.0.0
x402[fastapi,evm]>=2.1.0
sse-starlette>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
